+++ siteupdate/cplusplus/classes/DatacheckEntry/DatacheckEntry.h
@@ -1,3 +1,7 @@
+class Route;
+#include <list>
+#include <mutex>
+
 class DatacheckEntry
 {   /* This class encapsulates a datacheck log entry
 
@@ -55,42 +59,12 @@ class DatacheckEntry
 	bool fp;
 
 	static std::list<DatacheckEntry> errors;
-	static void add(Route *rte, std::string l1, std::string l2, std::string l3, std::string c, std::string i)
-	{	mtx.lock();
-		errors.emplace_back(rte, l1, l2, l3, c, i);
-		mtx.unlock();
-	}
+	static void add(Route*, std::string, std::string, std::string, std::string, std::string);
 
-	DatacheckEntry(Route *rte, std::string l1, std::string l2, std::string l3, std::string c, std::string i)
-	{	route = rte;
-		label1 = l1;
-		label2 = l2;
-		label3 = l3;
-		code = c;
-		info = i;
-		fp = 0;
-	}
+	DatacheckEntry(Route*, std::string, std::string, std::string, std::string, std::string);
 
-	bool match_except_info(std::array<std::string, 6> &fpentry)
-	{	// Check if the fpentry from the csv file matches in all fields
-		// except the info field
-		if (fpentry[0] != route->root)	return 0;
-		if (fpentry[1] != label1)	return 0;
-		if (fpentry[2] != label2)	return 0;
-		if (fpentry[3] != label3)	return 0;
-		if (fpentry[4] != code)		return 0;
-		return 1;
-	}
-
-	// Original "Python list" format unused. Using "CSV style" format instead.
-	std::string str()
-	{	return route->root + ";" + label1 + ";" + label2 + ";" + label3 + ";" + code + ";" + info;
-	}
-
-	bool operator < (DatacheckEntry &other)
-	{	return str() < other.str();
-	}
+	bool match_except_info(std::array<std::string, 6>&);
+	std::string str();
 };
 
-std::list<DatacheckEntry> DatacheckEntry::errors;
-std::mutex DatacheckEntry::mtx;
+bool operator < (DatacheckEntry &, DatacheckEntry &);
